refactor: add pages routing
